Re-add items to cache with a timeout to avoid limbo situation
Previously if the refresh task failed somehow (ie rabbit down, task
crash) then we were left in a limbo state where the cache indicated that
the job was running but it wasn't.  Hence the cache never got updated.

This change alters the way that we track whether a task is running.  We
now use a timeout value rather a special value.  Hence if the task fails
then it will be retried once the timeout expires.

Fixes #5