Optionally pass generated key over to async_refresh
The main reason for this is supporting the new Celery/Kombu serializer
where UUIDs don't raise anymore but are instead parsed as strs. This
change will generate two cache entries for jobs that iterables where
UUIDs get traversed as part of Job.hash due to the generated keys
being different in django-cacheback sync code and the async task (as
it will now receive a str instead of an UUID). The only key that
django-cacheback will be able to read is the first one and thus there
will always be a cache miss, a task executed and, seemingly, no return
value.

Link to the Kombu commit:
https://github.com/celery/kombu/commit/55452a15469cff38ad9a49e2730fe7250ec56ccd
Link to the Celery commit
:https://github.com/celery/celery/commit/6a9a540c37ec43257d587d8f1e9bea21ed588191

This will affect all versions of celery after 4.4.x when fetch_on_miss
is set to False.